Handle packages with differing amounts of trailing zeros
A new test demonstrates a case where publishing different builds of the
same package into multiple repos but with different version strings like
"1.0" and "1.0.0" gets into a situation where one of the builds become
invisible.

Address this by not assuming (in the spfs repository backend) that the
package's spfs tags are reliably normalized but instead check all the
possible forms that the version can take, to a reasonable limit of 5 parts:

- 1
- 1.0
- 1.0.0
- 1.0.0.0
- 1.0.0.0.0

Further changes to the code are needed to ensure that packages are saved
into the repo in a normalized format, but until such time that this happens
and all the problem packages in existing repos are fixed, this workaround
is necessary to avoid the problem of missing builds.

Fixes #966

Signed-off-by: J Robert Ray <jrray@jrray.org>