Fix default variant not contributing to build id
These two specs should produce a build with the same build id:

    pkg: demo/1.0.0
    build:
      options:
        - var: color/blue

    pkg: demo/1.0.0
    build:
      options:
        - var: color
      variants:
        - { color: blue }

They both build a package that has color=blue.

However, in the first case, the "color" option was not included in the
package's options, and did not contribute to the build id.

To handle this properly, `default_variants` needs to know when it is
returning the "default" variant. Therefore the BuildSpec no longer populates
`variants` with a default entry when no variants are defined.

Also, the override options (including host options) are needed to override any
values in the default options, otherwise filtering doesn't work as expected.

Signed-off-by: J Robert Ray <jrray@jrray.org>