Add version normalization
With these changes there are three ways to parse a version number and
render it back into a string:

1. verbatim: "1.0" remains "1.0"
2. normalized for display: "1.0" becomes "1.0.0"
3. normalized for storage: "1.0" becomes "1"

The default is (2), e.g., when using Version::to_string.

The verbatim case is used for preserving how versions were originally
written in recipes, for purposes of formatting error messages that
contain the original input, and for serializing the recipe and
preserving the original value.

The storage case is used for creating the spfs tag path, so that it
doesn't matter how many trailing zeros a version has in a recipe.

The normalized case is used in contexts where the verbatim case doesn't
apply, like in the output of `spk ls`, to make the output look more
uniform.

Closes #969

Signed-off-by: J Robert Ray <jrray@jrray.org>