Make SpfsRepository generic over the tag strategy
The purpose of this is to make it possible to create tests that write
content to the spfs repo using the original non-normalized strategy,
and confirm correct behavior when operating on a repo with mixed
content.

This commit does not add a way to open a repo with the old strategy.

Signed-off-by: J Robert Ray <jrray@jrray.org>