Framework for testing legacy-spk-version-testing
Add a new flavor of RepositoryHandle and rework the fixtures to make it
possible to create a repo configured to publish spk packages without
version normalization.

The new test utilizes this but that test in particular doesn't prove
much because `spk ls` would show the entry either way. Non-conforming
versions are not filtered out if this feature is not enabled.

Signed-off-by: J Robert Ray <jrray@jrray.org>