Test legacy-spk-version-tag behavior without flag enabled
In order to be able to test this behavior without having to run the test
suite twice, once with the feature enabled and once with it disabled,
make the related code available with or without the feature flag, and
make the behavior runtime configurable.

The default behavior is still set via the feature flag, but it can also
be enabled through a new hidden command line flag.

Change the demonstration test to not be conditional on the feature flag
being set.

Signed-off-by: J Robert Ray <jrray@jrray.org>