Use relative paths in documentation links
At SPI we host these docs in a sub folder of a larger documentation
site, and the non-relative links are broken since they expect these docs
to be hosted at the top level.

Some of the links were already relative, but this PR makes every "ref"
specify the path using a relative path in a consistent way, e.g., using
"./foo" in the hopes that refs added in the future will follow this example.

Signed-off-by: J Robert Ray <jrray@jrray.org>