Add feature to disable version normalization for writes
The new feature is called `legacy-spk-version-tags-for-writes`.

In order to safely migrate to version normalization in an environment
where the version of spk being used is not necessarily the latest
release everywhere, this feature causes spk to continue to publish
packages without version normalization. Local repo content will use
normalization, but not when writing to remote repos. This ensures older
versions of spk will still be able to read the published packages.

Once everywhere is using a sufficiently new version of spk, then this
feature can be turned off.

Since this is only a temporary problem at SPI, this code can be reverted
at some point in the future.

Another hidden command line flag has been added to the `publish` and
`export` subcommands to force this behavior even when the feature flag
is not enabled, for testing purposes. A new test has been added to
verify expected behavior on publish.

Signed-off-by: J Robert Ray <jrray@jrray.org>