Add more cases to legacy publish test
Make a distinction between how the version appears in the recipe versus
how it is written on the `spk publish` command line.

What ends up in the remote repository is expected to match what is
written in the recipe rather than what is given on the command line when
publishing.

Signed-off-by: J Robert Ray <jrray@jrray.org>