Change normalization scheme to have a minimum number of parts
Instead of normalizing "1.0.0" to "1", maintain a minimum size of three
parts in the interest of making the tag names more consistent and being
less of a disruptive change from the previous behavior.

An existing package using version "1.0.0" would not start being saved
to a different directory, avoiding potential backwards compatibility
issues.

Rename `Version::normalize` to `Version::strip_trailing_zeros` to avoid
confusion around what that method is expected to do and the fact that it now
returns something that doesn't match what a "normalized" version would
be now.

Signed-off-by: J Robert Ray <jrray@jrray.org>