Add flag to follow refs in info command
This is something that I regularly bump into, where I have a layer
digest, but want to see the contents of the manifest. Rather than
needing to copy and paste the digest into the info command again, this
will follow each child reference down to the manifest and print them
all.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>