Make spfs diff only check sizes on blobs
The main issue is the size of directories. If a path is committed on one
machine with a different filesystem type or with/without fuse, the
directory sizes captured into the manifest are platform-dependent, and
then rendering that manifest on a different system may cause `spfs diff`
to report changes.

One easy way to make this happen is to `spk build` a package while fuse
is enabled and then `spk env` the built package with fuse disabled.

```
$ spk env stdfs -- spfs diff
~/spfs/bin [size {0=>6}]
~/spfs/etc [size {1=>18}]
~/spfs/etc/spfs [size {1=>23}]
~/spfs/etc/spfs/startup.d [size {2=>53}]
~/spfs/lib [size {0=>6}]
~/spfs/spk [size {1=>17}]
~/spfs/spk/pkg [size {1=>19}]
~/spfs/spk/pkg/stdfs [size {1=>19}]
~/spfs/spk/pkg/stdfs/1.1.0 [size {1=>22}]
~/spfs/spk/pkg/stdfs/1.1.0/PF6B6TLF [size {5=>77}]
```

Signed-off-by: J Robert Ray <jrray@jrray.org>