Adds --stop-on-block/--stop option to solver (#1016)
This stops the solver when it first reaches a BLOCKED state and would
step back. This is useful when working out missing dependencies,
e.g. when porting a package between OSes or python versions.

* Add `--step-on-block/--step` option to solver (#1017)

This pauses the solver when a BLOCKED state is reached (a step-back
change). The solver will start up again when the user hits Enter.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>