Fix rpc error when encoding default is legacy
The rpc code was assuming that a default object header will be in
flatbuffers format, but at SPI we have changed it to default to legacy
until we are ready to adopt flatbuffer format files. Changing the
default causes rpc related tests to fail.

It is possible that rpc does not handle legacy format files without the change.

Since rpc now sends the FlatObject data over the wire as raw bytes, but
when using legacy format the header will be set to legacy, the receiving
end tries to respect the header and decode the raw bytes as if they are
in legacy format, as would be proper behavior when reading a legacy file
from disk. But when decoding an rpc message, the payload is _always_ in
flatbuffers format.

A new constructor was added that disregards the header's claimed format and
always treats the payload as flatbuffers encoded. This also means that
the server will write an object with a legacy header out to disk in
legacy format. Is this the desired behavior?

Should `graph::Object::new_with_default_header` be changed to stop
assuming what a "default" header is and hard code itself to use the new
encoding? It is unclear if the requirements of this method are to
produce a header set the new encoding, or to really just respect
whatever the default in the code is set to.

Signed-off-by: J Robert Ray <jrray@imageworks.com>