Fix `spk info` failing when using fuse
If inside an spfs runtime created via `spfs run <DIGEST> --` where
`<DIGEST>` is an object that does not exist in the local repository,
`spk info` would "crash" inside `env::current_env` with "Unknown Object"
while it looks for any solution data that might be present in an
annotation among the runtime layer stack.

The fundamental problem is that various spfs/spk command that are
invoked inside a runtime do not necessarily have access to the same set
of repositories that were used to create the runtime.

In particular, the `spfs info` and `spk info` commands will assume that
it will find any objects referenced by the runtime in the _local_
repository. Without using fuse, this can be okay because any objects end
up getting synced to the local repository when the runtime is created.
However, when using fuse, objects are not synced and may not be present
locally.

One possible way to solve this dilemma is to add a new field to the spfs
runtime to store the set of repos that were active when creating the
runtime. Then, operations that are interacting with the runtime
repository would have a way to know to look for objects in any of those
repos.

Signed-off-by: J Robert Ray <jrray@imageworks.com>