Require universal filesystem access for spfs-enter
On a modern 'fedora 40' linux distro, the enter process fails on any
attempt to modify the runtime or create directories in the repository in
my home directory. The error is EACCESS, which denotes that the process
doesn't have search permission on one of the directories involved. The
CAP_DAC_OVERRIDE privilege is needed in this configuraiton to allow the
enter process (running as root) full access to the file system.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>