Walk the path backwards when creating new directories
Our original algorithm walked down from the root, creating directories as
needed. This creates an operation of O(n) under all scenarios, where n
is the number of path components. However, we can generally assume
that these paths are going to be created once and then already exist
from that point forward. This mean that for a vast majority of the
operations, we are over-checking. This new system walks up from the full
path, checking each entry and tracking which do not exist. Once it
finds an existing parent, it walks back down creating each item. In a
worst case scenario, this becomes an O(2n) operation but changes our
best-case to an O(1). Eg if the path already exists, a single lstat call
confirms this and nothing else needs to be done. This feels like a
desirable trade-off given that most operations are likely to fall into
the "already exists" bucket.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>