Move the default distro compat values into config
Add a new spk config section for setting distro rules. Currently the
only rule is setting a compat value for a distro, but others may come
later (like renaming the distro, e.g., "rocky" -> "rhel9", for more
flexibility).

An invalid compat value in the config will produce an error at the point
where that value is needed, rather than attempting to parse the compat
value when parsing the config. This could be changed later, but for now
it prevents a bad config file from competely breaking all spk commands.

Signed-off-by: J Robert Ray <jrray@jrray.org>