Handle when multiple packages embed the same package
The solver was panicking with "state not found" in the case where it
encounters a package that embeds some other package, but that other
package has already been embedded by some third package. It wanted to
"eject" the real package (again) but no "real" package was in the
solution at that point.

Now it checks that the package it wants to "eject" is a not an embedded
package. This situation was given its own `IncompatibleReason` so the
solver output will detail the packages involved.

Signed-off-by: J Robert Ray <jrray@jrray.org>