Add a heartbeat between spfs-monitor and spfs-fuse
As mentioned in #895, there are cases where spfs-fuse never shuts down.
An easy way to repro this is to `kill -9` the spfs-monitor process, so
it doesn't get a chance to clean up the runtime normally.

We observe spfs-fuse processes accumulating over time on our CI runners.
It could be from users canceling pipelines (maybe the runner does a
`kill -9` of all the child processes?). It is also extremely common for
zombie runtimes to accumulate over time (not just on CI runners), perhaps
spfs-monitor is crashing / failing to cleanup commonly?

This is a proof of concept and enabling this or tuning the timing of the
heartbeat and timeout should be a configurable thing but these are hard
coded for now.

Signed-off-by: J Robert Ray <jrray@jrray.org>