Make sync progress bars optional
The goal is to have the ability for commands like `spfs run` and `spk
env` to not show a progress bar when running the Syncer. A new
`--no-progress-bars` flag has been added where appropriate.

The Syncer type is no longer generic over the SyncReporter in order for
Sync::get_syncer to be able to return a Syncer with either reporter type.
The existing code was not doing anything useful with it being generic
over the SyncReporter.

Some `*_with_reporter` alternate functions have been added to spk-exec
since these functions were hard-coded to show progress bars.

Signed-off-by: J Robert Ray <jrray@jrray.org>