Rename AsVersion to AsVersionIdent
Since there was also an existing `as_version` method on the Ident types,
change these to use the trait instead. If `as_version` could be confused
for something that returns a Version, then by extension the other
methods on Ident should be similarly renamed. This also keeps the naming
convention consistent.

Signed-off-by: J Robert Ray <jrray@jrray.org>