Make environment variables to preserve configurable
We were preserving $TMPDIR specfically in a hard-coded way and there was
no way to opt out of it if desired. Now it has come to light that
$LD_LIBRARY_PATH is also cleared when entering an spfs runtime. We would
like to preserve that value too. Add a new configuration option so each
site can decide to opt in to the preserving. Doing so may come with
risks of privilege escalation.

Signed-off-by: J Robert Ray <jrray@jrray.org>