Move Repository::address to dedicated trait
This makes it possible for other things to have a trait bound on
`Address` instead of all of `Repository`, which means other types can
implement `Address` without having to also implement all of
`Repository` (not shown in this commit).

Also change the method signature to return `Cow` since there are cases
where an existing `Url` value can be borrowed.

Signed-off-by: J Robert Ray <jrray@jrray.org>