Combine FsRepository and OpenFsRepository
Use the type state pattern to combine these two types into a single
type. This is some groundwork for creating new types to represent repos
that have renders or not. This code doesn't make much use of the type
state pattern yet.

`FsRepository` used to represent a repo that may not be opened yet, this
is now represented by the type alias:

    type MaybeOpenFsRepository = FsRepository<MaybeOpenFsRepositoryImpl>

An already opened repository has the same name, but is now a type alias:

    type OpenFsRepository = FsRepository<OpenFsRepositoryImpl>

The original types have been renamed to these *Impl names:

    FsRepository -> MaybeOpenFsRepositoryImpl
    OpenFsRepository -> OpenFsRepositoryImpl

Signed-off-by: J Robert Ray <jrray@jrray.org>