De-duplicate code in renderer_win.rs and renderer_unix.rs
Some of the differences between these two files appear to be minor
changes that were done on the unix side and not also done on the windows
side, which is a good example of why it is bad to have this code
duplicated in two places.

It is also frustrating to be working in linux and push changes only to
discover from CI that the changes break the windows build.

Now the functions that only exist (or are implemented) on the linux side
live in the renderer_unix.rs file, but most of it lives as a shared
implementation in renderer.rs.

Signed-off-by: J Robert Ray <jrray@jrray.org>