Refactor rendering logic into separate methods
This change was prompted by wanting to refactor the code to work towards
having different trait bounds for the functions that need a repository
that supports renders (e.g., rendering into a repository) versus
functions that render into a plain directory.

Rework `RenderType` so the method that renders into a hardlink doesn't
need to worry about getting a `RenderType::Copy` value.

Signed-off-by: J Robert Ray <jrray@jrray.org>