Refactor FsRepository's renders
The goal is to make it so it is possible to open repo without creating
the renders directory for the current user, in situations where that
renders directory will not be used. Examples include when using fuse or
when running `spfs clean`.

This introduces different flavors of a "render store" besides
`RenderStore` that either represent a render store that hasn't been
created yet (but can be) or the lack of a render store. Operations that
don't require a render store can access the local repository without the
renders directory being created.

Signed-off-by: J Robert Ray <jrray@jrray.org>