Preliminary work to adopt Resolvo as the solver
Resolvo is a CDCL sat solver for package managers and will hopefully
prove to be much better at solving quickly than our mostly brute-force
approach.

Making a checkpoint here after achieving a simple "solve" to get a feel
for how spk type map onto Resolvo's concepts. Although currently the
smallest unit the solver works with is a package, and that likely
needs to change to a component.

Signed-off-by: J Robert Ray <jrray@jrray.org>