Crudely support global vars
Resolvo doesn't offer a way to declare constraints that aren't
associated with a package. The strategy employed here is to create
synthetic packages to represent global variables. Whenever a new
possible value for a global variable is discovered, the solver needs to
be restarted so when "candidates" for a global variable are requested,
it returns a "package" (solvable) for each possible value that variable
might take. Then, the requirement for packages that have that specific
value can be models as a dependency on the synthetic package.

It is not too common for packages to have install requirements on
non-namespaced vars, so in practice this solver restart may not be a
frequent occurrence. Also, top level options (not yet implemented) will
allow packages that expect vars with other values to be excluded instead
of having to ingest those different values.

Of course it would be preferable for the solver to support this concept
without the restart mechanic, but maybe it is worth considering making
it a requirement for packages to only have namespaced var requirements.

Another option is to maintain a cache of known var values somehow either
in the repo or in an external service. This could cut down on the number
of instances where the solver needs to be restarted.

Signed-off-by: J Robert Ray <jrray@jrray.org>