Bundle tag_path with build information
This is extremely useful for code using spk as a library when iterating
over packages/builds to be able to also learn what the spfs tag backing
the build was. Since some packages may live in the repo with
non-normalized tags (created in the before times), it is otherwise a lot
of extra work to figure out what the spfs tag would be for a build
returned by list_package_builds.

Signed-off-by: J Robert Ray <jrray@imageworks.com>