Finish integrating cdcl solver into existing solver tests
Not all solver tests are applicable. Where possible, for tests the cdcl
solver doesn't pass yet, add #[should_panic] so the test suite still
passes as a whole. As the cdcl solver can pass more tests, this will
highlight the tests that can then be enabled for the cdcl solver.

Signed-off-by: J Robert Ray <jrray@jrray.org>