Pass the "All" component test
This required rethinking the type backing NameIds to subdivide a package
into its components. For example, "python:run" and "python:build" are
considered separate packages by the solver.

It's not possible for a single request to resolve to multiple
resolvables. Either a request for "All" would have to be expanded into
requests for the individual components, or there has to be a [virtual]
package "python:all" that represents a dependency on all the components
of that package. Since it is not possible to know the full set of
components for a request like "python:all", where the set of components
may differ between the different versions of python, then only the
latter approach is feasible.

Dependencies are needed to ensure the solve only contains components from the
same spk package build.

Signed-off-by: J Robert Ray <jrray@jrray.org>