Improve sorting to avoid embedded packages
At SPI the newest python version available on centos is an embedded
package, but it is undesirable for `spk env python` to choose an embedded
python package.

It also follows that it would have preferred to build from source a
newer version of a package instead of use an existing build of an older
version.

Signed-off-by: J Robert Ray <jrray@imageworks.com>