Add a replacement for running solve through formatter
The abstract solver now has a `run_and_print_resolve` that takes a
formatter, which makes it possible to run either the old or new solver
in a consistent way.

The original `run_and_print_resolve` function has become `pub(crate)` to
catch all the places that needed to be modified to use the new function.

Signed-off-by: J Robert Ray <jrray@imageworks.com>
Signed-off-by: J Robert Ray <jrray@jrray.org>