spfs: add an option for mounting overlayfs layers using syscalls
Teach spfs to mount "/spfs" using direct mount syscalls.
This is an opt-in feature so that users can test this alongside the
current "mount"-based implementation.

Unfortunately, the path length limits that are still present when using
the "lowerdir" option with older kernels. We autodetect when "lowerdir+"
is available to match what we do when using "mount" commands.

Related-to: #968
Related-to: #1164