Fix resolvo solver with migration-to-components
The solver tests were not passing if this feature is enabled, and this
exposed a few problems in not handling ':all' requests correctly. When
generating the "VersionSet" for dependencies, the PkgRequest needs to
have the correct component populated. The logic for skipping or not
skipping candidates when build from source needed to handle ':all' and
it was restructured for clarity. When filtering candidates for a request
of a component of a specific build, source builds needed to be filtered
out.

Signed-off-by: J Robert Ray <jrray@imageworks.com>