Solve `spk info filename.spk.yaml` failing with file not found
At least for the simple case where no workspace file is present, fix
`find_package_template_by_file` so it compares with canonicalized path
names.

Change `Workspace::load_or_default` so that if the `--workspace`
argument is a directory (that doesn't contain a workspace.spk.yaml file)
it still roots the virtual workspace to the given directory. This allows
for workspace tests to load a virtual workspace in a temporary
directory.

Add a very basic cmd_view test that verifies it is possible to use `spk
view --variants filename.spk.yaml`, something which is required by the
CI automation at SPI.

Signed-off-by: J Robert Ray <jrray@imageworks.com>