Add tests to demonstrate workspace problems
The new workspace feature has introduced a variety of regressions for
using `spk info --variants <filename>` and will error with a file not
found error for files that do actually exist.

One problem is that when a workspace spec doesn't exist, then the only
possible files that can be inspected have to exist in the root of the
virtual workspace generated, ignoring that the command line contains the
name of a valid file.

Another problem is when the workspace does exist, then the only possible
files that can be inspected have to be defined in the workspace, even
when the command line contains the name of a valid file.

Perhaps the design can be reconsidered to not use a "virtual workspace"
in these cases or at least have a fallback that tries to open the file
directly.

Signed-off-by: J Robert Ray <jrray@jrray.org>