Replace workspace Result with enum with Result+Error
Also stop exiting on workspace load error

It is bad form for library code to exit the process. It also makes the
test suite die early when attempting to write tests that exercise the
error path.

All the uses of `must_be_found` were at the application level and could
be changed to return a generic miette error instead of exiting.

Signed-off-by: J Robert Ray <jrray@jrray.org>

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>