Make git source tests not require a git clone
If using a different vcs tool besides git to check out the spk project,
these tests would fail, as they assume the working copy is a git clone.

Modify the tests to create a new git clone in a temporary directory so
they don't depend on the working copy being a git clone.

Signed-off-by: J Robert Ray <jrray@jrray.org>