Switch to dynamically loading spec files into workspaces
This allows commands to work as they were before where absolute and
deeply nested file paths can still be given even when a workspace is
present. We can decide to warn or refine this behavior in the future if
it makes sense, but maintaining the existing behavior makes more sense
for now

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>

Also Add tests to demonstrate workspace problems

The new workspace feature has introduced a variety of regressions for
using `spk info --variants <filename>` and will error with a file not
found error for files that do actually exist.

One problem is that when a workspace spec doesn't exist, then the only
possible files that can be inspected have to exist in the root of the
virtual workspace generated, ignoring that the command line contains the
name of a valid file.

Another problem is when the workspace does exist, then the only possible
files that can be inspected have to be defined in the workspace, even
when the command line contains the name of a valid file.

Perhaps the design can be reconsidered to not use a "virtual workspace"
in these cases or at least have a fallback that tries to open the file
directly.

Signed-off-by: J Robert Ray <jrray@jrray.org>