Add configure_for_build_environment to the Solver trait
Fix commented out code in cmd_test for build stage tests and add a new
test that checks the basic behavior of a build stage test that the test
environment includes the build dependencies of the package being tested.

Lift the old solver's configure_for_build_environment method to
the Solver trait. Although the other test stages have code that only lives
in the cmd_test crate to configure the solver properly, the build stage
uses this method that was defined in the solver itself. While the logic
of configure_for_build_environment could be duplicated in cmd_test, it
is also reused by other code in the solver so it made sense to preserve
the logic as a reusable solver method.

Signed-off-by: J Robert Ray <jrray@jrray.org>