Skip invalid template files when loading workspace
For the use case of `spk info --variants <valid spec file>`, this
command would fail with an error about a different file if another file
exists with a name matching `*.spk.yaml` but loading this other file
fails for any reason.

Like #1190 and #1191, this command is expected to succeed and is not
intended to interact with any other files other than the one named on
the command line.

Signed-off-by: J Robert Ray <jrray@jrray.org>