Remove legacy-spk-version-tags-for-write feature
This feature was intended to only exist temporarily to support old
non-normalized tags while phasing out old builds of spk that couldn't
handle normalized tags.

Now at SPI with this feature still enabled, we're unintentionally
continuing to create new non-normalized content in the repo.

Remove this feature so nothing is still creating non-normalized tags.
Repositories are always opened in "normalized mode" now so the type
distinction has gone away.

One test had to be deleted because it relied on the ability to write
non-normalized tags, which is no longer possible with this change.

The other related feature for reading non-normalized content will be
removed in a future PR.

Signed-off-by: J Robert Ray <jrray@jrray.org>