Fix `spfs clean` failing if a tag namespace exists
This slipped through testing, the clean would immediately fail when it
encounters a directory named like "test#ns" as it fails to parse the
tag spec, reporting that the "org" is invalid.

This error is avoided by allowing "#ns" to be the suffix of an org and
then `spfs clean` seems to work as expected, but ignores all the cases
where it _shouldn't_ accept an "org" that contains a namespace.

The current tag stream code doesn't really understand tag namespaces or
offer a way to choose to walk into them or skip them. It was intended
that `spfs clean` would visit all tags in all namespaces, but otherwise
other tag operations would only see tags in the configured namespace.

Signed-off-by: J Robert Ray <jrray@jrray.org>