Restructure spk-solve for multiple solvers
Move `Solver` into a nested module and rename it publicly as
`StepSolver` in order to make way for other solver implementations and
for the name `Solver` to become available for repurposing it later.

The name `StepSolver` was inspired by how the solver makes progress by
calling the `step_state` function.

The `solver_test` module has been separated from the `solver` module
since this is destined to become a module of tests that are generic over
the solver implementation.

Signed-off-by: J Robert Ray <jrray@jrray.org>