Add Resolvo examples to the Solver Usage documentation
A number of the examples are not really applicable to Resolvo.

Some recipes were added to be able to recreate the output from the
existing examples (though the existing output is a little out of date).

The tabs and tab shortcodes were not rendering code blocks as theme
documentation suggests it will, so it has been modified to wrap the
"content" in a code block. That required removing the triple backticks
from the markdown file. Perhaps this can be improved?

Signed-off-by: J Robert Ray <jrray@jrray.org>