Cleaner now walks all tags in all namespaces
Make Cleaner "tag namespace aware" instead of expecting that the tag
iterator returns tags across all namespaces. It now iterates over the
existing tag namespaces in order to visit all tags in all namespaces.

Add another clean test that sets a tag namespace.

Have coverage to a scenario where the Cleaner is called with a repo that
has a tag namespace set on it. The Cleaner should still operate on the
whole repo and all namespaces when looking for strong references to
objects.

Signed-off-by: J Robert Ray <jrray@jrray.org>