Revert the original "#ns" suffix hack
Now that the Cleaner is "tag namespace aware", the fs tag iterator does
not need to return tags that belong to another namespace. It does
however need to recognize and skip files that represent tags in another
namespace. But now it will once again impossible to name a tag with a
path that would put it in a different namespace.

Signed-off-by: J Robert Ray <jrray@jrray.org>