Limit tag namespace names to not have subdirectories
Now that `spfs clean` only discovers and walks tags in tag namespaces
defined at the top level, it is unsafe to allow creating tag namespaces
in subdirectories.

It is assumed that this change will not affect any existing uses of tag
namespaces and that a deprecation process is not necessary.

Signed-off-by: J Robert Ray <jrray@jrray.org>