Update to Rust 1.87.0
Clippy is mad about a number of large enums in this new version. Adding
Box in targeted enum variants to avoid the lint while also avoiding
putting a Box in a position that gets in the way of pattern matching
conveniently.

Making use of `cargo +nightly rustc -- -Zprint-type-sizes` to see enum
variant sizes.

Signed-off-by: J Robert Ray <jrray@jrray.org>