Improve Debug representation of Digest
Flatbuffers generates a Debug representation of Digest that looks like
this:

    Digest {
        bytes: [
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0,
            0
        ]
    }

It is painful when turning on debug output when the log has many pages
of this kind of output. The debug output of an spfs runtime, which
includes the list of layer digests, can be thousands of lines long for
large environments with many layers.

This change makes the Debug representation of Digest look like this:

    Digest("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA====")

The list of bytes is changed to our standard base32 representation.

To make this happen, the flatbuffers-generated code has to be modified
after it is generated. The tool ast-grep is tailored to do this kind of
transformation. To make that tool available, it is "cargo installed" in
the contexts it is needed. Like how flatc is required to build the
project, ast-grep will also be required.

Signed-off-by: J Robert Ray <jrray@jrray.org>
Signed-off-by: J Robert Ray <jrray@imageworks.com>